https://github.com/opensourceBIM/BIMserver/issues/1001

+++ PluginBase/src/org/bimserver/emf/IdEObject.java
@@ -20,7 +20,7 @@ package org.bimserver.emf;
 import org.eclipse.emf.ecore.EObject;
 
 public interface IdEObject extends EObject {
-	int getExpressId();
+	long getExpressId();
 	long getOid();
 	int getRid();
 	int getPid();
