+++ PluginBase/src/org/bimserver/plugins/serializers/EmfSerializer.java
@@ -39,7 +39,7 @@ public abstract class EmfSerializer implements Serializer, StreamingReader {
 	private Mode mode = Mode.HEADER;
 	private ProjectInfo projectInfo;
 	private boolean normalizeOids;
-	private int expressIdCounter = 1;
+	private long expressIdCounter = 1;
 
 	protected static enum Mode {
 		HEADER, BODY, FOOTER, FINISHED
@@ -71,7 +71,7 @@ public abstract class EmfSerializer implements Serializer, StreamingReader {
 		this.mode = mode;
 	}
 
-	protected int getExpressId(IdEObject object) {
+	protected long getExpressId(IdEObject object) {
 		if (normalizeOids && object.getExpressId() == -1) {
 			((IdEObjectImpl)object).setExpressId(expressIdCounter ++);
 		}
